refactor: split session/context.rs into SRP sub-modules (50-line limit)
CodeTether-Provenance-ID: ctprov_f25ff1facb194d8ea9219b2f0441b938
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: 26895400-3387-476c-b440-859864603fac
